修正
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@1224 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/net/l1j/server/clientpackets/C_CreateChar.java
@@ -150,7 +150,7 @@ public class C_CreateChar extends ClientBasePacket {
 		L1ClassFeature classFeature = L1ClassFeature.newClassFeature(pc.getType());
 		short initHp = (short)classFeature.InitHp();
 		short initMp = (short)classFeature.InitMp(pc.getWis());
-		short initLucky = (short)classFeature.InitLucky();
+		int initLucky = classFeature.InitLucky();
 		int [] spawn = classFeature.InitSpawn(pc.getType());
 		pc.setId(IdFactory.getInstance().nextId());
 		pc.setClassId(classFeature.InitSex(pc.get_sex()));
